pyproject.toml
[project]
name = "actualai"
version = "0.1.0"
description = ""
authors = [
    {name = "Pop101"}
]
readme = "README.md"
requires-python = "<4.0,>=3.12"
dependencies = [
    "actualpy (>=0.15,<0.17.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "duckduckgo-search (>=8.0.0,<9.0.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "langchain-ollama (>=0.3.1,<0.4.0)",
    "python-levenshtein (>=0.27.1,<0.28.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
